@@ -17,31 +17,31 @@
     "compare-versions": "^3.5.1",
     "dateformat": "^3.0.3",
     "deepmerge": "^4.1.1",
-    "expo": "^35.0.0",
-    "expo-analytics-segment": "~7.0.0",
-    "expo-av": "~7.0.0",
-    "expo-blur": "~7.0.0",
-    "expo-constants": "~7.0.0",
-    "expo-file-system": "~7.0.0",
-    "expo-image-manipulator": "~7.0.0",
-    "expo-image-picker": "~7.0.0",
-    "expo-intent-launcher": "^8.0.0",
-    "expo-linear-gradient": "~7.0.0",
-    "expo-localization": "~7.0.0",
-    "expo-location": "~7.0.0",
-    "expo-media-library": "~7.0.0",
-    "expo-permissions": "~7.0.0",
-    "expo-secure-store": "~7.0.0",
-    "expo-web-browser": "~7.0.0",
+    "expo": "^36.0.0",
+    "expo-analytics-segment": "~8.0.0",
+    "expo-av": "~8.0.0",
+    "expo-blur": "~8.0.0",
+    "expo-constants": "~8.0.0",
+    "expo-file-system": "~8.0.0",
+    "expo-image-manipulator": "~8.0.0",
+    "expo-image-picker": "~8.0.1",
+    "expo-intent-launcher": "~8.0.0",
+    "expo-linear-gradient": "~8.0.0",
+    "expo-localization": "~8.0.0",
+    "expo-location": "~8.0.0",
+    "expo-media-library": "~8.0.0",
+    "expo-permissions": "~8.0.0",
+    "expo-secure-store": "~8.0.0",
+    "expo-web-browser": "~8.0.0",
     "i18n-js": "^3.3.0",
     "jwt-decode": "^2.2.0",
     "lodash": "^4.17.15",
     "lottie-react-native": "~2.6.1",
     "native-base": "^2.13.8",
     "prop-types": "^15.7.2",
-    "react": "16.8.3",
+    "react": "16.9.0",
     "react-fast-compare": "^2.0.4",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.1.tar.gz",
     "react-native-animatable": "^1.3.3",
     "react-native-animated-swiper": "^0.0.23",
     "react-native-autogrow-textinput": "^5.2.0",
@@ -52,28 +52,28 @@
     "react-native-elements": "^1.2.6",
     "react-native-expo-image-cache": "^4.0.8",
     "react-native-flash-message": "^0.1.15",
-    "react-native-gesture-handler": "~1.3.0",
+    "react-native-gesture-handler": "~1.5.0",
     "react-native-htmlview": "^0.15.0",
     "react-native-input-scroll-view": "^1.9.5",
-    "react-native-maps": "~0.25.0",
+    "react-native-maps": "0.26.1",
     "react-native-masked-text": "^1.13.0",
     "react-native-modal-datetime-picker": "^7.6.0",
     "react-native-multiple-select": "^0.5.3",
     "react-native-paper": "^2.16.0",
     "react-native-phone-input": "^0.2.4",
-    "react-native-reanimated": "~1.2.0",
+    "react-native-reanimated": "~1.4.0",
     "react-native-responsive-screen": "^1.3.0",
     "react-native-sectioned-multi-select": "^0.7.2",
     "react-native-side-menu": "^1.1.3",
     "react-native-slider": "^0.11.0",
     "react-native-slowlog": "^1.0.2",
     "react-native-snap-carousel": "^3.8.2",
-    "react-native-svg": "9.9.2",
+    "react-native-svg": "9.13.3",
     "react-native-swipeout": "^2.3.6",
     "react-native-swiper": "^1.5.14",
     "react-native-switch-toggle": "^1.1.2",
     "react-native-vector-icons": "^6.6.0",
-    "react-native-webview": "7.0.5",
+    "react-native-webview": "7.4.3",
     "react-navigation": "^4.0.10",
     "react-navigation-fluid-transitions": "^0.3.2",
     "react-navigation-stack": "^1.9.4",
@@ -90,7 +90,7 @@
     "victory-native": "^33.0.1"
   },
   "devDependencies": {
-    "babel-preset-expo": "^7.0.0",
+    "babel-preset-expo": "^8.0.0",
     "eslint-config-prettier": "^6.7.0",
     "eslint-plugin-prettier": "^3.1.1",
     "react-native-svg-transformer": "^0.13.0",
