@@ -16,7 +16,7 @@ export default class KeyboardShift extends Component {
     shift: new Animated.Value(0)
   };
 
-  componentWillMount() {
+  componentDidMount() {
     this.keyboardDidShowSub = Keyboard.addListener(
       "keyboardDidShow",
       this.handleKeyboardDidShow
