@@ -71,6 +71,47 @@ export default GoogleSEAPreview = props => {
           style={[styles.headlineRowContainer, dynamicStyle(language).arabic]}
         >
           <View style={[styles.headlineRow, dynamicStyle(language).arabic]}>
+            {details && (
+              <Text
+                style={[
+                  styles.headline,
+                  dynamicStyle(language).arabic,
+                  inputH1
+                    ? GlobalStyles.orangeTextColor
+                    : GlobalStyles.darkGrayTextColor
+                ]}
+              >
+                {translate("Headline")} {translate(headlineNums[0])}
+              </Text>
+            )}
+
+            <Text style={[styles.headlineText, dynamicStyle(language).arabic]}>
+              {headline1
+                ? headlineArray[0] + " "
+                : translate("Add") + " " + translate("Headline")}
+              |{" "}
+              <Text
+                style={[styles.headlineText, dynamicStyle(language).arabic]}
+              >
+                {headline2
+                  ? headlineArray[1] + " "
+                  : translate("Add") + " " + translate("Headline")}
+              </Text>
+              |{" "}
+              <Text
+                style={[styles.headlineText, dynamicStyle(language).arabic]}
+              >
+                {headline3
+                  ? headlineArray[2] + " "
+                  : translate("Add") + " " + translate("Headline")}
+              </Text>
+            </Text>
+          </View>
+        </View>
+        <View
+          style={[styles.headlineRowContainer, dynamicStyle(language).arabic]}
+        >
+          <View style={[styles.headlineRow, dynamicStyle(language).arabic]}>
             <View style={[styles.headerContent, dynamicStyle(language).arabic]}>
               {details && (
                 <Text
@@ -164,8 +205,6 @@ export default GoogleSEAPreview = props => {
           </View>
         </View>
       </View>
-
-      <View style={styles.descriptionGrayLine} />
       <View
         style={[
           styles.headersCol,
@@ -198,9 +237,16 @@ export default GoogleSEAPreview = props => {
               " " +
               translate("Description") +
               " " +
-              translate("1")}
+              translate("1") +
+              " "}
+          {description2 ||
+            translate("Add") +
+              " " +
+              translate("Description") +
+              " " +
+              translate("2")}
         </Text>
-        <Text
+        {/* <Text
           style={[
             styles.headlineText,
             styles.descriptionText,
@@ -215,6 +261,7 @@ export default GoogleSEAPreview = props => {
               " " +
               translate("2")}
         </Text>
+       */}
       </View>
     </View>
   );
