@@ -1,5 +1,5 @@
 import React, { Component } from "react";
-import { Text, View, Modal, StyleSheet } from "react-native";
+import { Text, View, StyleSheet } from "react-native";
 import { SafeAreaView } from "react-navigation";
 import { BlurView } from "expo-blur";
 import Header from "../../MiniComponents/Header";
@@ -8,7 +8,7 @@ import InputField from "../../MiniComponents/InputField";
 import EmailTransparentIcon from "../../../assets/SVGs/EmailTransparent";
 import CustomButtons from "../../MiniComponents/CustomButtons";
 import FlashMessage from "react-native-flash-message";
-
+import Modal from "react-native-modal";
 export default class CSVModal extends Component {
   state = { email: "", emailError: "", incomplete: false };
 
@@ -61,9 +61,17 @@ export default class CSVModal extends Component {
     const { translate } = screenProps;
 
     return (
-      <Modal animationType="fade" transparent visible={isVisible}>
+      <Modal
+        style={{ margin: 0 }}
+        animationIn="fadeIn"
+        animationOut="fadeOut"
+        isVisible={isVisible}
+      >
         <BlurView intensity={100} tint="dark" style={{ height: "100%" }}>
-          <SafeAreaView style={{ justifyContent: "space-evenly" }}>
+          <SafeAreaView
+            style={{ justifyContent: "space-evenly" }}
+            forceInset={{ top: "always", bottom: "always" }}
+          >
             <Header
               closeButton={true}
               actionButton={this.actionButton}
@@ -108,5 +116,5 @@ export default class CSVModal extends Component {
 }
 
 const CSVStyle = StyleSheet.create({
-  customButtonStyle: { height: "13%", width: "70%" }
+  customButtonStyle: { height: 50, width: "70%" }
 });
