@@ -5,12 +5,14 @@ import formatNumber from "../../../formatNumber";
 import PlaceholderLine from "../../../MiniComponents/PlaceholderLine";
 import globalStyles from "../../../../GlobalStyles";
 import ImpressionsIcons from "../../../../assets/SVGs/CampaignCards/ImpressionsIcon";
+import SwipeUpsIcon from "../../../../assets/SVGs/CampaignCards/SwipeUpsIcon";
 import SpendIcons from "../../../../assets/SVGs/SpendingIcon";
 import FrequencyIcon from "../../../../assets/SVGs/CampaignDetail/FrequencyIcon";
 import CPIIcon from "../../../../assets/SVGs/CampaignDetail/CPIIcon";
 import CPVIcon from "../../../../assets/SVGs/CampaignDetail/CPVIcon";
 import ReachIcon from "../../../../assets/SVGs/CampaignDetail/ReachIcon";
 import { heightPercentageToDP } from "react-native-responsive-screen";
+import { Small } from "../../../MiniComponents/StyledComponents";
 
 export default class SingleMetric extends Component {
   render() {
@@ -32,6 +34,9 @@ export default class SingleMetric extends Component {
       case "Reach":
         Icon = ReachIcon;
         break;
+      case "swipeup rate":
+        Icon = SwipeUpsIcon;
+        break;
       default:
       case "eCPV":
         Icon = CPVIcon;
@@ -48,20 +53,41 @@ export default class SingleMetric extends Component {
           />
         )}
         {!loadingCampaignStats ? (
-          <Text style={[styles.title]}>
-            {translate(metric)}
-            {"\n "}
-            <Text
-              style={[styles.numbers, { fontFamily: "montserrat-regular" }]}
-            >
-              {formatNumber(
-                Number.isInteger(metricValue)
-                  ? metricValue
-                  : parseFloat(metricValue).toFixed(2),
-                !metric.toLowerCase().includes("c")
+          <View>
+            <Text style={[styles.title]}>{translate(metric)}</Text>
+            <View style={globalStyles.row}>
+              {metric.toLowerCase().includes("c") && (
+                <Small
+                  style={[
+                    styles.numbers,
+                    { fontSize: 9, fontFamily: "montserrat-regular" }
+                  ]}
+                >
+                  $
+                </Small>
               )}
-            </Text>
-          </Text>
+              <Text
+                style={[styles.numbers, { fontFamily: "montserrat-regular" }]}
+              >
+                {formatNumber(
+                  Number.isInteger(metricValue)
+                    ? metricValue
+                    : parseFloat(metricValue).toFixed(2),
+                  true
+                )}
+              </Text>
+              {metric.toLowerCase() === "swipeup rate" && (
+                <Small
+                  style={[
+                    styles.numbers,
+                    { fontSize: 9, fontFamily: "montserrat-regular" }
+                  ]}
+                >
+                  %
+                </Small>
+              )}
+            </View>
+          </View>
         ) : (
           <PlaceholderLine width={70} />
         )}
