@@ -180,6 +180,7 @@ class CampaignDetails extends Component {
     this.setState(prevState => ({
       expand: !prevState.expand
     }));
+    this.scroll.scrollTo({ x: 0, y: 0, animated: true });
     this.toggle();
   };
 
@@ -447,10 +448,7 @@ class CampaignDetails extends Component {
                   textAlign: "left",
                   fontSize: 15,
                   paddingTop: 3,
-                  alignSelf: "center",
-                  justifyContent: "center",
-                  flex: 1,
-                  alignItems: "center"
+                  flex: 1
                 }}
                 campaignStatus={loading ? null : selectedCampaign.status}
               />
@@ -544,6 +542,7 @@ class CampaignDetails extends Component {
               <ScrollView
                 contentContainerStyle={{ height: hp(100) }}
                 scrollEnabled={!this.state.expand}
+                ref={ref => (this.scroll = ref)}
                 style={{ maxHeight: "100%" }}
               >
                 <View style={[styles.mainCard]}>
