@@ -56,7 +56,7 @@ const styles = StyleSheet.create({
   linkText: {
     fontFamily: "montserrat-regular",
     color: "#197132",
-    paddingLeft: 0
+    paddingLeft: 0,
   },
   descriptionGrayLine: {
     borderTopColor: "#EDEDED",
