@@ -2,18 +2,21 @@ import React from "react";
 import { TouchableOpacity, View } from "react-native";
 import { Text } from "native-base";
 import styles from "./styles";
+import { Small } from "../../../MiniComponents/StyledComponents";
+import GradientButton from "../../../MiniComponents/GradientButton";
 
+const kFormatter = num => {
+  return Math.abs(num) > 999
+    ? (Math.abs(num) / 1000).toFixed(1) + "k"
+    : Math.abs(num).toFixed(Number.isInteger(num) ? 0 : 2);
+};
 export default props => {
   let { content, selected, onPressFunction } = props;
-  kFormatter = num => {
-    return Math.abs(num) > 999
-      ? (Math.abs(num) / 1000).toFixed(1) + "k"
-      : Math.abs(num).toFixed(2);
-  };
   return (
-    <TouchableOpacity
-      rounded
-      onPress={() => onPressFunction(!selected ? content : {})}
+    <GradientButton
+      gradientDirection="horizontal"
+      onPressAction={() => onPressFunction(!selected ? content : {})}
+      transparent={!selected}
       style={selected ? styles.filledButton : styles.emptyButton}
     >
       <View style={styles.keywordRowOuterView}>
@@ -26,19 +29,25 @@ export default props => {
         </Text>
         <View style={styles.keywordRowView}>
           <View style={styles.displayStatsView}>
-            <Text uppercase style={[styles.numbercontentStyle, {}]}>
-              {this.kFormatter(content.spend)}
+            <Small style={{ fontSize: 8 }}>$</Small>
+            <Text
+              adjustsFontSizeToFit={true}
+              uppercase
+              style={[styles.numbercontentStyle, {}]}
+            >
+              {kFormatter(content.spend)}
             </Text>
           </View>
           <View style={styles.displayStatsView}>
             <Text
               uppercase
+              numberOfLines={2}
               style={[
                 styles.numbercontentStyle,
                 { textAlign: "center", display: "flex" }
               ]}
             >
-              {this.kFormatter(content.clicks)}
+              {kFormatter(content.clicks)}
             </Text>
           </View>
           <View style={styles.displayStatsView}>
@@ -48,9 +57,10 @@ export default props => {
             >
               {content.ctr}
             </Text>
+            <Small style={{ fontSize: 9 }}>%</Small>
           </View>
         </View>
       </View>
-    </TouchableOpacity>
+    </GradientButton>
   );
 };
