@@ -464,11 +464,21 @@ export const get_google_campiagn_details = (
             type: actionTypes.SET_GOOGLE_CAMPAIGN_STATS,
             payload: { loading: false, data: data }
           });
-        else
+        else {
+          let endDate = new Date(data.campaign.end_time);
+          endDate.setDate(endDate.getDate() + 2);
+          if (
+            !data.campaign.completed &&
+            data.campaign.status === "REMOVED" &&
+            endDate < new Date()
+          ) {
+            dispatch(refundGoogleCampaignAmount(data.campaign.id));
+          }
           return dispatch({
             type: actionTypes.SET_CAMPAIGN,
             payload: { loading: false, data: data }
           });
+        }
       })
       .catch(err => {
         showMessage({
@@ -819,5 +829,23 @@ export const downloadGoogleCSV = (campaign_id, email, showModalMessage) => {
         if (data.message) showModalMessage(data.message, "success");
       })
       .catch(err => showModalMessage(err));
+  };
+};
+
+export const refundGoogleCampaignAmount = campaign_id => {
+  return (dispatch, getState) => {
+    GoogleBackendURL()
+      .get(`request/refund/`, {
+        businessid: getState().account.mainBusiness.businessid,
+        id: campaign_id
+      })
+      .then(res => res.data)
+      .then(data => {
+        if (data.message)
+          showMessage({ message: data.message, type: "success" });
+      })
+      .catch(err => {
+        errorMessageHandler(err);
+      });
   };
 };
